fix(client): use correct length for fwmarks
Corrects the length of the fwmark attribute, which should be uint32
instead of uint16.

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/net/netfilter/ipvs/ip_vs_ctl.c?h=v5.10#n2974

Fixes #1

+++ client.go
@@ -36,7 +36,7 @@ type Service struct {
 	Timeout   uint32
 	Flags     Flags
 	Port      uint16
-	FWMark    uint16
+	FWMark    uint32
 	Family    AddressFamily
 	Protocol  Protocol
 }
